+++ packages/studio/src/components/renders/RenderQueueItem.tsx
@@ -1,4 +1,5 @@
-import { memo, useCallback, useState, useEffect, useRef } from "react";
+import { memo, useCallback, useState } from "react";
+import { VideoFrameThumbnail } from "../ui/VideoFrameThumbnail";
 import type { RenderJob } from "./useRenderQueue";
 
 interface RenderQueueItemProps {
@@ -20,53 +21,6 @@ function formatTimeAgo(timestamp: number): string {
 }
 
 /** Static frame extracted once via hidden video + canvas. */
-function RenderThumbnail({ src }: { src: string }) {
-  const [frame, setFrame] = useState<string | null>(null);
-  const didExtract = useRef(false);
-
-  useEffect(() => {
-    if (didExtract.current) return;
-    didExtract.current = true;
-
-    const video = document.createElement("video");
-    video.crossOrigin = "anonymous";
-    video.muted = true;
-    video.preload = "metadata";
-
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d");
-
-    const cleanup = () => {
-      video.src = "";
-      video.load();
-    };
-
-    video.addEventListener("loadedmetadata", () => {
-      video.currentTime = Math.min(2, video.duration * 0.1 || 2);
-    });
-
-    video.addEventListener("seeked", () => {
-      if (!ctx) return;
-      canvas.width = video.videoWidth;
-      canvas.height = video.videoHeight;
-      ctx.drawImage(video, 0, 0);
-      setFrame(canvas.toDataURL("image/jpeg", 0.7));
-      cleanup();
-    });
-
-    video.addEventListener("error", cleanup);
-    video.src = src;
-    video.load();
-
-    return cleanup;
-  }, [src]);
-
-  if (!frame) {
-    return <div className="w-full h-full bg-neutral-800 animate-pulse rounded" />;
-  }
-
-  return <img src={frame} alt="" draggable={false} className="w-full h-full object-cover" />;
-}
 
 export const RenderQueueItem = memo(function RenderQueueItem({
   job,
@@ -125,7 +79,7 @@ export const RenderQueueItem = memo(function RenderQueueItem({
                 className="absolute inset-0 transition-opacity duration-150"
                 style={{ opacity: hovered ? 0 : 1 }}
               >
-                <RenderThumbnail src={viewSrc} />
+                <VideoFrameThumbnail src={viewSrc} />
               </div>
             </>
           )}
